Refactoring printResults out of bin for reuse
